Rename pkg to maxdimassociator
