+++ pkg/job/maxdimassociator/associator_ecs_test.go
@@ -1,4 +1,4 @@
-package associator
+package maxdimassociator
 
 import (
 	"testing"
