+++ pkg/job/maxdimassociator/associator_globalaccelerator_test.go
@@ -1,4 +1,4 @@
-package associator
+package maxdimassociator
 
 import (
 	"testing"
