+++ pkg/job/maxdimassociator/associator_lambda_test.go
@@ -1,4 +1,4 @@
-package associator
+package maxdimassociator
 
 import (
 	"testing"
